+++ data-raw/HYDAT_internal_data/process_internal_data.R
@@ -10,9 +10,10 @@ library(tidyverse)
 
 #' A tibble of all Canadian Stations stations and their names.
 allstations <- realtime_stations() %>%
+  mutate(HYD_STATUS = "ACTIVE", REAL_TIME = TRUE) %>%
   bind_rows(hy_stations()) %>%
   distinct(STATION_NUMBER, .keep_all = TRUE) %>%
-  select(STATION_NUMBER, STATION_NAME, PROV_TERR_STATE_LOC, LATITUDE, LONGITUDE) 
+  select(STATION_NUMBER, STATION_NAME, PROV_TERR_STATE_LOC, HYD_STATUS, REAL_TIME, LATITUDE, LONGITUDE) 
 
 use_data(allstations, overwrite = TRUE)